[USB3]: HID update, HID driver now support USB Tablet [USB3]: HID now finally got the ability to report mouse coordinates to kernel [Kernel]: Bugs fixed in serial and kernel screen print function, where negative values were not printed previously

+++ Kernel/stdio.cpp
@@ -20,7 +20,7 @@ static char* chars = "0123456789ABCDEF";
 
 extern void debug_print(const char *text, ...);
 
-char* sztoa(size_t value, char* str, int base)
+char* sztoa(int64_t value, char* str, int base)
 {
 	if (base < 2 || base > 16)
 		return nullptr;
